Public API Reference

Publicly Exported Functions

WaveguideModes.lookup_rwg — Function

lookup_rwg(wgspec::AbstractString) -> (a, b)

Find the dimensions of a rectangular waveguide given its "spec" (a string using standard waveguide nomenclature).

Input Argument

wgspec::AbstractString: A standard (EIA, RCSC, or IEC) abbreviation for a rectangular waveguide size. Examples include "WR650" (EIA), "WG7" (RCSC), and "R22" (IEC).

Return Value

(a, b): Waveguide inner dimensions in the x and y directions, respectively, expressed as Unitful quantities with units of u"inch".

WaveguideModes.rwg_modes — Function

rwg_modes(a, b, f; nmodes=10, ϵᵣ=1.0, tanδ=0.0, σ=Inf*u"S/m", Rq=0.0u"m") -> modedata

Compute cutoff frequency, guide wavelength, and attenuation constant for first few modes of a rectangular waveguide.

Note: This function is intended for programmatic use. For interactive use, see rwg_modetable.

Required Positional Arguments

a, b: The waveguide inner dimensions as any Unitful length quantity, e.g. b=0.8128u"cm", b=320u"mil", or b=0.320u"inch". a and b must be expressed in the same length units.

f: The frequency as any Unitful frequency quantity, e.g. f=3u"GHz", f=3e9u"Hz" or f=3000u"MHz".

Optional Keyword Arguments

nmodes: The number of modes to be treated. Defaults to 10.

ϵᵣ or epsr: The dielectric constant for the material filling the waveguide.

tanδ or tandel: The loss tangent for the dielectric material filling the waveguide.

σ or sigma: Bulk conductivity of the metal walls in any Unitful quantity with dimensions equal to those of u"S/m". Defaults to Inf*u"S/m" (perfect electric conductor).

Rq: Surface RMS roughness of the metal walls in any Unitful length dimension, e.g. 6u"μm". Defaults to 0u"m" (perfectly smooth). The surface roughness is used with σ to compute a (lower) effective conductivity via the MetalSurfaceImpedance package.

Return Value

modedata::Matrix{Any}: A matrix of size (nmodes, 6), containing in each row [p, m, n, fco, λg, α] for a single mode. Here p = "TE" for a TE mode or "TM" for a TM mode, m and n are integer mode indices in the x and y directions, resp., fco is the cutoff frequency expressed in the same units as input argument f (but without attached units), λg is the guide wavelength expressed in the same units as a and b (but without attached units), and α is the mode attenuation constant in units of dB/unitlength (without attached units), where unitlength is one unit of length in the same units as a and b. The modes are listed in order of increasing cutoff frequency. For cutoff modes (other than TE10 or TE01 which are calculated more accurately) the value of λg will be Inf.

WaveguideModes.rwg_modetable — Method

rwg_modetable(wgspec, frequency; kwargs...)

Pretty-print a table of rectangular waveguide mode properties: cutoff frequency, guide wavelength, and attenuation constant.

Guide wavelength and attenuation for the TE₁₀ and TE₀₁ modes are accurately computed at frequencies below, at, or above cutoff, for both smooth and rough imperfectly conducting surfaces, using the Gradient method as detailed in the paper: "Analytical waveguide model precisely predicting loss and delay including surface roughness," IEEE Trans. Microwave Theory Tech., vol. 66, no. 6 (2018): pp. 2649-2662. For other modes the standard perturbational formulas are used in conjunction with the Gradient method to determine effective surface impedance. For cutoff modes (other than TE₁₀ and TE₀₁) the value of guide wavelength will be printed out as Inf.

Note: This function is intended for interactive use. For programmatic use, see rwg_modes. The table is printed to the user's console using the PrettyTables package.

Required Positional Arguments

wgspec::AbstractString: A standard (EIA, RCSC, or IEC) abbreviation for a rectangular waveguide size. Examples include "WR650" (EIA), "WG7" (RCSC), and "R22" (IEC).

frequency: The frequency expressed as a Unitful quantity with dimensions of inverse time. Examples: 2.35u"GHz", 2350u"MHz", 2.35e9u"Hz", etc.

Optional Keyword Arguments

length_unit:: A Unitful:Units object with dimension of length. Default value is u"inch". The table will print values in the "guide wavelength" column using this unit; similarly the attenuation column will use units of dB/length_unit.

ϵᵣ or epsr: Dielectric constant of material filling the waveguide. Default value is 1.0.

tanδ or tandel: Loss tangent of material filling the waveguide. Default value is 0.0.

σ or sigma: The bulk conductivity of the waveguide metal walls expressed as a Unitful quantity with dimensions the same as those of u"S/m". Default value is Inf*u"S/m".

Rq: The RMS surface roughness of the waveguide walls expressed as a Unitful quantity having dimension of length. Default value is 0.0u"m".

col_fmts::String = ["%s", "%i", "%i", "%#.8g", "%8.4f", "%8.4f"]: A vector of C-style format strings used to format the six columns of the table.

WaveguideModes.rwg_modetable — Method

rwg_modetable(a, b, frequency; kwargs...)

An alternative method for rwg_modetable allowing explicit specification of the waveguide dimensions. Here, instead of wgspec, the first two arguments are the x and y dimensions of the waveguide, expressed as Unitful quantities having dimension of length. Both must be expressed in the same units, e.g. u"mm" or u"inch". The default value of length_unit for this method is whatever length unit a and b use.

WaveguideModes.rwgte10gz — Function

rwgte10gz(a, b, f; ϵᵣ, tanδ, σ, Rq) -> (γ, Z)

Accurately compute γ and Z (prop. constant and wave impedance) of TE10 mode in rectangular guide with rough walls.

Required Positional Arguments

a, b: The waveguide x and y dimensions, Unitful quantities with dimensions of length.

f: Frequency, a Unitful quantity with dimensions of inverse time.

Optional Keyword Arguments

ϵᵣ, tanδ: Dielectric constant and loss tangent, respectively, of material filling the waveguide. Default values are 1.0 and 0.0, respectively.

σ: The bulk conductivity of the waveguide metal walls, a Unitful quantity with the same dimensions as u"S/m". Default value is Inf*u"S/m".

Rq: The RMS surface roughness of the waveguide walls, a Unitful quantity with dimension of length. Default value is 0.0u"m"

Return Values

γ: Complex propagation constant [neper/m]

Z: Complex wave impedance [Ω]

References

[1] Lomakin, Konstantin, Gerald Gold, and Klaus Helmreich. "Analytical waveguide model precisely predicting loss and delay including surface roughness." IEEE Transactions on Microwave Theory and Techniques 66, no. 6 (2018): 2649-2662.

[2] Gold, Gerald, and Klaus Helmreich. "A physical surface roughness model and its applications." IEEE Transactions on Microwave Theory and Techniques 65, no. 10 (2017): 3720-3732.

[3] Lomakin, Konstantin, Gerald Gold, and Klaus Helmreich. "Transmission line model for rectangular waveguides accurately incorporating loss effects." In 2017 IEEE 21st Workshop on Signal and Power Integrity (SPI), pp. 1-4. IEEE, 2017.

rwgte10gz(wgspec, f; kwargs...) -> (γ, Z)

This method accepts a rectangular waveguide "specification string" instead of a and b dimensions.

Required Positional Inputs

wgspec::AbstractString: A standard (EIA, RCSC, or IEC) abbreviation for a rectangular waveguide size. Examples include "WR650" (EIA), "WG7" (RCSC), and "R22" (IEC).

Optional Keyword Arguments

ϵᵣ, tanδ: Dielectric constant and loss tangent, respectively, of material filling the waveguide. Default values are 1.0 and 0.0, respectively.

σ: The bulk conductivity of the waveguide metal walls, a Unitful quantity with the same dimensions as u"S/m". Default value is Inf*u"S/m".

Rq: The RMS surface roughness of the waveguide walls, a Unitful quantity with dimension of length. Default value is 0.0u"m"

Return Values

γ: Complex propagation constant [neper/m]

Z: Complex wave impedance [Ω]

rwgte10gz(a::Real, b::Real, f::Real; ϵᵣ::Real=1.0, tanδ::Real=0.0, σ::Real=Inf, Rq::Real=0.0) -> (γ, Z)
rwgte10gz(a::Real, b::Real, f::Real; epsr::Real=1.0, tandel::Real=0.0, sigma::Real=Inf, Rq::Real=0.0) -> (γ, Z)

This method accepts SI values as pure numbers (without attached units).

Required Positional Arguments

a, b: The waveguide x and y dimensions [m].

f: Frequency [Hz]

Optional Keyword Arguments

ϵᵣ or epsr: Dielectric constant of material filling the waveguide. Default value is 1.0.

tanδ or tandel: Loss tangent of material filling the waveguide. Default value is 0.0.

σ or sigma: The bulk conductivity of the waveguide metal walls [S/m]. Default value is Inf.

Rq: The RMS surface roughness of the waveguide walls [m]. Default value is 0.0.

Return Value

The tuple (γ, Z) where:

γ: Complex propagation constant [neper/m]

Z: Complex wave impedance [Ω]

Publicly Exported Types

WaveguideModes.RWG — Type

RWG <: HomogeneousMetallicWaveguide

A struct representing a section of metallic, rectangular waveguide, uniformly filled with dielectric. It is assumed that the (possibly non-smooth) metal walls are either perfectly conducting or form a "good conductor", in the sense that the standard perturbational formulas for losses in waveguide walls hold for higher-order modes. For the dominant TE10 and the TE01 modes, accurate formulas from the recent literature are used. See the docstring for function rwgte10gz for more information.

Fields

a::Float64: The x-dimension of the waveguide [m].

b::Float64: The y-dimension of the waveguide [m].

l::Float64: The length of the waveguide section (along z) [m].

ϵᵣ::Float64: The dielectric constant for the material filling the guide [unitless].

tanδ::Float64: The loss tangent for the material filling the guide [unitless].

σ::Float54: The bulk conductivity of the waveguide metal walls [S/m].

Rq::Float64: The metal wall RMS surface roughness [m].

modes::Vector{RWGMode}: The modes treated in this guide. If provided, the list should be sorted in order of increasing cutoff frequency.

Both positional and keyword argument constructors for RWG are available.

WaveguideModes.RWGMode — Type

RWGMode <: HomogeneousMetalPipeMode

Struct representing an electromagnetic mode of a rectangular waveguide, homogeneously filled with dielectric.

Fields:

p::TETM: Mode type.

m::Int: Index for modal variation along x direction.

n::Int: Index for modal variation along y direction.

kco::Float64: Cutoff wavenumber [radian/meter].

f::Float64: Frequency [Hz].

γ::ComplexF64: Attenuation constant [neper/meter].

Z::ComplexF64: Modal wave impedance normalized to η₀, the impedance of free space [unitless].

Non-Exported API (Subject to Change)

Private Functions

WaveguideModes.alphaploss — Method

alphaploss(params, k, Rsnorm)

Compute the attenuation constant of a rectangular waveguide using the classical power loss method.

Input Arguments

params: A NamedTuple or other destructurable object containing the following fields:

a, b: waveguide dimensions in x and y directions [m].

m, n, p::TETM: Mode indices and mode type.

kco: Cutoff wavenumber of mode [rad/m].

k: Wavenumber in the dielectric filling the waveguide.

Rsnorm: Ratio of surface resistance of the waveguide metal walls to the intrinsic impedance of the dielectric filling the waveguide.

Return Value

α: For frequencies above cutoff, the attenuation constant [np/m] calculated via the classical power loss method. For frequencies below cutoff, zero is returned.

References:

R. E. Collin, Field Theory of Guided Waves 2cd Edition, IEEE Press, 1991, Table 5.2, p. 351.

J. Uher, J. Bornemann and U. Rosenberg, Waveguide Components for Antenna Feed Systems: Theory and CAD, Artech House, 1991, p. 114.

WaveguideModes.alphaploss — Method

alphaploss(rwg::RWG, modeindex, k, Rsnorm)

Compute the attenuation constant of a rectangular waveguide using the classical power loss method.

Input Arguments

rwg: An RWG instance with modes instantiated.

modeindex: The index into rwg.modes for the mode of interest.

k: Wavenumber in the dielectric filling the waveguide.

Rsnorm: Ratio of surface resistance of the waveguide metal walls to the intrinsic impedance of the dielectric filling the waveguide.

Return Value

α: For frequencies above cutoff, the attenuation constant [np/m] calculated via the classical power loss method. For frequencies below cutoff, zero.

References:

R. E. Collin, Field Theory of Guided Waves 2cd Edition, IEEE Press, 1991, Table 5.2, p. 351.

J. Uher, J. Bornemann and U. Rosenberg, Waveguide Components for Antenna Feed Systems: Theory and CAD, Artech House, 1991, p. 114.

WaveguideModes.collin_gamma — Method

collin_gamma(a, b, m, n, p, k, Zsnorm)

Compute variational estimate of the complex attenuation constant for a rectangular waveguide mode, using the variational method described on pages 350-354 of the reference.

Arguments

a, b: Waveguide dimensions along x and y, respectively [m].

m, n: Nonnegative integer mode numbers along x and y, respectively, not both zero, and neither zero if p == TM.

p::TETM: Mode type (TE or TM).

k: The wavenumber in the dielectric medium filling the waveguide.

Zsnorm: The surface impedance of the metal waveguide walls normalized to the intrinsic impedance of the dielectric medium.

Return Value

γ: A variational estimate of the complex attenuation constant [Np/m] of the specified mode.

WaveguideModes.collin_gamma_objective — Method

collin_gamma_objective(Δγari, parameters)

Objective function for collin_gamma.

Arguments

Δγari: A 2-vector containing real and imaginary parts of Δγa. The latter is defined such that the current estimate for γ is obtained as im*β + Δγa / a. Note that Δγa is unitless, and is also very small for a waveguide with good conductor walls.

parameters: A named tuple consisting of at least the following fields:

a: Waveguide x dimension [m].

p::TETM: Desired mode type

β: Phase constant of the degenerate modes assuming PEC walls [rad/m].

β²: The square of β.

f11, term11b, A12, A21, term22b: Constants needed to evaluate the matrix elements defined in Collin Eq. (66) on page 353. These have all been multiplied by waveguide dimension a to render them unitless.

Return Value

TBC

WaveguideModes.collin_gammas — Method

collin_gammas(a, b, m, n, k, Zsnorm)

Compute variational estimate of the complex attenuation constant for a rectangular waveguide mode, using the variational method described on pages 350-354 of the reference.

Arguments

a, b: Waveguide dimensions along x and y, respectively [m].

m, n: Nonnegative integer mode numbers along x and y, respectively, not both zero, and neither zero if p == TM.

k: The wavenumber in the dielectric medium filling the waveguide.

Zsnorm: The surface impedance of the metal waveguide walls normalized to the intrinsic impedance of the dielectric medium.

Return Value

(; γte, γtm): TE and TM complex attenuation constants [Np/m].

WaveguideModes.collinmatrix — Method

collinmatrix(γ, parameters)

Compute the 2x2 matrix from Eq. (66) on page 353 of Collin (after each row multiplied by a to render unitless).

Arguments

γ: The complex attenuation constant [Np/m] from which the matrix is to be computed.

parameters: A named tuple consisting of at least the following fields:

β²: The square of β.

f11, term11b, A12, A21, term22b: Constants needed to evaluate the matrix elements defined in Collin Eq. (66) on page 353. These have all been multiplied by waveguide dimension a to render them unitless.

Return Value

detabs2: The magnitude squared of the determinant of the matrix equation in Eq. (66), after each row has been multiplied by a to make the matrix entries unitless.

WaveguideModes.compute_As_kxakybs — Method

compute_As_kxakybs(FGHz, rwg, mode_params; iterate=false)

Compute dispersion matrices for a rectangular waveguide with lossy walls using 4 choices of (δx, δy)

Input Arguments

FGHz: The desired analysis frequency in GHz.

rwg::RWG: An RWG instance with at least fields a, b, σ, and Rq set.

mode_params: Named tuple or struct with fields m::Integer, n::Integer, p::TETM defining the desired mode.

iterate: If false (default), then the δx and δy values computed from [1] are used directly to compute kx and ky and thus γ. If true, then these values are used as starting values in a root-finding algorithm implementing the numerical search described in [2].

Return value

(As, kxakybs), where

As: A 4-vector containing four complex matrices of size (4,2) containing the coefficients of E0 and η*H0 in the modal dispersion relation. As[i] contains the matrix obtained by choosing (δx, δy) to be

i==1: (δx1, δy1)

i==2: (δx2, δy1)

i==1: (δx1, δy2)

i==1: (δx2, δy2)

kxakybs: A 4-vector containing tuples (kx*a, ky*b) computed for the same four cases as As.

References

[1] Yeap, Kim Ho, et al. "Attenuation in circular and rectangular waveguides." Electromagnetics 37.3 (2017): 171-184.

[2] Yeap, Kim Ho, et al. "Attenuation in rectangular waveguides with finite conductivity walls." Radioengineering 20.2 (2011).

WaveguideModes.cwgkcoa — Method

cwgkcoa(p::TETM, m::Integer, n::Integer)

Compute the the product of cutoff wavenumber times radius for a circular waveguide.

Required Positional Arguments

p: Mode type (TE or TM).

m: Mode number for azimuthal variation. Must be nonnegative.

n: Mode number for radial variation. Must be positive.

Return Value

kcoa: The cutoff wavenumber of the mode times the waveguide radius.

WaveguideModes.cwgkza_abe — Method

cwgkza_abe(p::TETM, m, umn, ka, Zsn)

Compute product of z-directed complex wavenumber kz and waveguide radius a for a circular waveguide.

Arguments

p::TETM: Mode type (TE or TM).

m::Integer: Mode index for variation in ϕ direction.

umn::Float64: Equal to kco * a, where kco is the cutoff wavenumber for the mode of interest. (For circular guides this is a zero of the Bessel function J_m or a zero of its derivative).

ka::ComplexF64: Product of wavenumber (for the dielectric filling the guide) times the waveguide radius.

Zsn::ComplexF64: The ratio of the metal wall surface impedance to the intrinsic impedance of the dielectric material filling the waveguide.

Return Value

'kza::Complex64': Product of the waveguide radius with the z-directed wavenumber [1/m] including both propagation (imaginary part) and attenuation (real part).

Reference

T. Abe and Y. Yamaguchi, "Propagation Constant Below Cutoff Frequency in a Circular Waveguide with Conducting Medium", IEEE Trans. MTT, Vol. MTT-29, no. 7, July 1981, pp. 707-712.

WaveguideModes.cwgkza_ploss — Method

cwgkza_ploss(p::TETM, m, umn, ka, Zsn)

Compute product of z-directed complex wavenumber kz and waveguide radius a for a circular waveguide. Estimate the attenuation for modes above cutoff using the power loss method.

Arguments

p::TETM: Mode type (TE or TM).

m::Integer: Mode index for variation in ϕ direction.

umn::Float64: Equal to kco * a, where kco is the cutoff wavenumber for the mode of interest. (For circular guides this is a zero of the Bessel function J_m or a zero of its derivative).

ka::ComplexF64: Product of wavenumber (for the dielectric filling the guide) times the waveguide radius.

Zsn::ComplexF64: The ratio of the metal wall surface impedance to the intrinsic impedance of the dielectric material filling the waveguide.

Return Value

'kza::Complex64': Product of the waveguide radius with the z-directed wavenumber [1/m] including both propagation (imaginary part) and attenuation (real part).

Reference

R. Collin, Field Theory of Guided Waves, 2cd Edition, Table 5.5.

WaveguideModes.cwgkza_yeap — Method

cwgkza_yeap(p::TETM, m, umn, ka, Zsn)

Compute product of z-directed complex wavenumber kz and waveguide radius a for a circular waveguide.

Arguments

p::TETM: Mode type (TE or TM).

m::Integer: Mode index for variation in ϕ direction.

umn::Float64: Equal to kco * a, where kco is the cutoff wavenumber for the mode of interest. (For circular guides this is a zero of the Bessel function J_m or a zero of its derivative).

ka::ComplexF64: Product of wavenumber (for the dielectric filling the guide) times the waveguide radius.

Zsn::ComplexF64: The ratio of the metal wall surface impedance to the intrinsic impedance of the dielectric material filling the waveguide.

Return Value

'kza::Complex64': Product of the waveguide radius with the z-directed wavenumber [1/m] including both propagation (imaginary part) and attenuation (real part).

Reference

K. H. Yeap et al: "Attenuation in Circular and Rectangular Waveguides", Electromagnetics, Vol. 37, No. 3, 2017, pp. 171-184

WaveguideModes.e0h0mat — Method

e0h0mat(kxa, kyb, params)

Compute the dispersion matrix A whose null space defines the possible z-component field coefficients.

Input Arguments

kxa, kyb: The (unitless) product of proposed x and y modal wavenumbers times the x and y dimensions of the waveguide, resp.

params: A struct or named tuple containing the following parameters of the waveguide problem:

m, n: Mode numbers in the x and y directions, resp.

a, b: Waveguide x and y dimensions, resp. [m]

k: Wavenumber [radians/m] in the dielectric filling the waveguide.

ηwn: The wall surface impedance normalized to the intrinsic impedance of the dielectric filling the waveguide.

Return Value

A: ComplexF64 matrix of dimensions 4×2 computed using the provided inputs, whose nullspace spans the possible modal vectors [E₀, η*H₀]. Here E₀ is the coefficient of the z-component of electric field, H₀ is the coefficient of the z-component of magnetic field, and η is the intrinsic impedance of the dielectric filling the waveguide.

WaveguideModes.is_html_environment — Method

is_html_environment()

Returns true if code is running under IJulia or Pluto, and false otherwise.

WaveguideModes.mysqrt — Method

mysqrt(x)

Same as sqrt unless sqrt(x) is pure negative imaginary in which case it returns -sqrt(x) (i.e., positive pure imaginary).

WaveguideModes.nlsolvedkxadkyb — Method

nlsolvedkxadkyb(Δs, params) -> residual::SVector{4, Float64}

Function defining a NonLinearProblem whose solution provides rigorously correct kx and ky values for a rectangular waveguide.

Positional Arguments

Δs: A 4-vector containing the unitless values [real(Δkxa), imag(Δkxa), real(Δkyb), imag(Δkyb)]. HereΔkxais the departure ofkxfrom its nominal value (for PEC walls) ofmπ/a. Similarly,Δkybis the departure ofkyfrom its nominal value ofnπ/b. The entries inΔs` are to be adjusted to produce a zero return vector.

params: A named tuple containing the fixed parameters of the waveguide problem. These are

m and n: Mode indices in the x and y directions, respectively.

k: The (complex) wavenumber in the dielectric filling the waveguide [radians/m].

a and b: The waveguide (inner) dimensions in the x and y directions, respectively [m].

ηwn: Waveguide wall surface impedance normalized to the intrinsic impedance of the dielectric filling the waveguide.

Return Value

residual: A static 4-vector containing the real and imaginary parts of the two expressions that must be zeroed for a valid solution.

Reference:

WaveguideModes.nlsolvekxakyb — Method

nlsolvekxakyb(rikakbs, params) -> residual::SVector{4, Float64}

Function defining a NonLinearProblem whose zero solves the determinantal equations for the kx and ky values of a rectangular waveguide.

Positional Arguments

rikakbs: A 4-vector containing the unitless values [real(kxa), imag(kxa), real(kyb), imag(kyb)]. Herekxais the product ofkxanda(the waveguide x dimension) Similarly,kybis the product ofkyandb(the waveguide y dimension). The entries inrikakbs` are to be adjusted to produce a zero return vector.

params: A named tuple containing the fixed parameters of the waveguide problem. These are

m and n: Mode indices in the x and y directions, respectively.

k: The (complex) wavenumber in the dielectric filling the waveguide [radians/m].

a and b: The waveguide (inner) dimensions in the x and y directions, respectively [m].

ηwn: Waveguide wall surface impedance normalized to the intrinsic impedance of the dielectric filling the waveguide.

Return Value

residual: A static 4-vector containing the real and imaginary parts of the two expressions that must be zeroed for a valid solution.

Reference:

[1] Yeap, Kim Ho, et al. "Attenuation in circular and rectangular waveguides." Electromagnetics 37.3 (2017): 171-184. See Eqs. (6a) and (6b). [2] P. Simon, "Notes on Attenuation in Circular and Rectangular Waveguides", Eqs (7a) and (7b).

WaveguideModes.rwggz — Method

rwggz(p, m, n, a, b, f; ϵᵣ=1.0, tanδ=0.0, σ=Inf, Rq=0.0) -> (γ, Z)
rwggz(p, m, n, a, b, f; epsr=1.0, tandel=0.0, sigma=Inf, Rq=0.0) -> (γ, Z)

Accurately compute γ and Z (prop. constant and wave impedance) of a mode in rectangular guide with rough walls.

This method accepts SI values as pure numbers (without attached units).

Required Positional Arguments

p::TETM: The mode type.

m::Int, n::Int: Mode indices in the x and y directions, resp.

a::Real, b::Real: The waveguide x and y dimensions [m].

f: Frequency [Hz].

Optional Keyword Arguments

ϵᵣ or epsr: Dielectric constant of material filling the waveguide. Default value is 1.0.

tanδ or tandel: Loss tangent of material filling the waveguide. Default value is 0.0.

σ or sigma: The bulk conductivity of the waveguide metal walls [S/m]. Default value is Inf.

Rq: The RMS surface roughness of the waveguide walls [m]. Default value is 0.0.

Return Value

The tuple (γ, Z) where:

γ: Complex propagation constant [neper/m]

Z: Complex wave impedance [Ω]

References:

[1] Yeap, Kim Ho et al., "Attenuation in circular and rectangular waveguides." Electromagnetics 37, no. 3 (2017): 171-184.

WaveguideModes.rwgkco — Method

rwgkco(a, b, m, n)

Compute the cutoff frequency for a rectangular waveguide.

Input Variables

a, b: The dimensions of the waveguide in the x and y directions, respectively.

m, n: Mode indices in the x and y directions, respectively.

Return value:

kco: Cutoff wavenumber in inverse units of a and b.

WaveguideModes.update — Method

update(mode::HomogeneousMetalPipeMode; f::Real, γ::Complex, Z::Complex) -> mode

Create a copy of an existing mode with updated frequency-dependent fields and return the updated mode.

Private Types

WaveguideModes.HomogeneousMetalPipeMode — Type

abstract type HomogeneousMetalPipeMode <: MetalPipeMode end

Electromagnetic mode supported by homogeneously filled metal pipes. Concrete subtypes should contain at least the following fields:

p::TETM: Mode classification (TE or TM).

f::Float64: Frequency [Hz].

γ::ComplexF64: Frequency-dependent attenuation constant [np/m].

Z::ComplexF64: Frequency-dependent wave impedance [Ω].

WaveguideModes.MetalPipeMode — Type

abstract type MetalPipeMode <: Mode end

Electromagnetic mode supported by metal pipes, possibly filled with inhomogeneous dielectric.

WaveguideModes.Mode — Type

abstract type Mode end

Arbitrary electromagnetic mode. Should contain the following fields:

γ::ComplexF64: Attenuation constant in z direction [np/m]

This page was generated using Literate.jl.