Public API Reference

Publicly Exported Functions

WaveguideModes.lookup_rwg — Function

lookup_rwg(wgspec::AbstractString) -> (a, b)

Find the dimensions of a rectangular waveguide given its "spec" (a string using standard waveguide nomenclature).

Input Argument

wgspec::AbstractString: A standard (EIA, RCSC, or IEC) abbreviation for a rectangular waveguide size. Examples include "WR650" (EIA), "WG7" (RCSC), and "R22" (IEC).

Return Value

(a, b): Waveguide inner dimensions in the x and y directions, respectively, expressed as Unitful quantities with units of u"inch".

WaveguideModes.rwg_modes — Function

rwg_modes(a, b, f; nmodes=10, ϵᵣ=1.0, tanδ=0.0, σ=Inf*u"S/m", Rq=0.0u"m") -> modedata

Compute cutoff frequency, guide wavelength, and attenuation constant for first few modes of a rectangular waveguide.

Note: This function is intended for programmatic use. For interactive use, see rwg_modetable.

Required Positional Arguments

a, b: The waveguide inner dimensions as any Unitful length quantity, e.g. b=0.8128u"cm", b=320u"mil", or b=0.320u"inch". a and b must be expressed in the same length units.

f: The frequency as any Unitful frequency quantity, e.g. f=3u"GHz", f=3e9u"Hz" or f=3000u"MHz".

Optional Keyword Arguments

nmodes: The number of modes to be treated. Defaults to 10.

ϵᵣ or epsr: The dielectric constant for the material filling the waveguide.

tanδ or tandel: The loss tangent for the dielectric material filling the waveguide.

σ or sigma: Bulk conductivity of the metal walls in any Unitful quantity with dimensions equal to those of u"S/m". Defaults to Inf*u"S/m" (perfect electric conductor).

Rq: Surface RMS roughness of the metal walls in any Unitful length dimension, e.g. 6u"μm". Defaults to 0u"m" (perfectly smooth). The surface roughness is used with σ to compute a (lower) effective conductivity via the MetalSurfaceImpedance package.

Return Value

modedata::Matrix{Any}: A matrix of size (nmodes, 6), containing in each row [p, m, n, fco, λg, α] for a single mode. Here p = "TE" for a TE mode or "TM" for a TM mode, m and n are integer mode indices in the x and y directions, resp., fco is the cutoff frequency expressed in the same units as input argument f (but without attached units), λg is the guide wavelength expressed in the same units as a and b (but without attached units), and α is the mode attenuation constant in units of dB/unitlength (without attached units), where unitlength is one unit of length in the same units as a and b. The modes are listed in order of increasing cutoff frequency. For cutoff modes (other than TE10 or TE01 which are calculated more accurately) the value of λg will be Inf.

WaveguideModes.rwg_modetable — Method

rwg_modetable(wgspec, frequency; kwargs...)

Pretty-print a table of rectangular waveguide mode properties: cutoff frequency, guide wavelength, and attenuation constant.

Guide wavelength and attenuation for the TE₁₀ and TE₀₁ modes are accurately computed at frequencies below, at, or above cutoff, for both smooth and rough imperfectly conducting surfaces, using the Gradient method as detailed in the paper: "Analytical waveguide model precisely predicting loss and delay including surface roughness," IEEE Trans. Microwave Theory Tech., vol. 66, no. 6 (2018): pp. 2649-2662. For other modes the standard perturbational formulas are used in conjunction with the Gradient method to determine effective surface impedance. For cutoff modes (other than TE₁₀ and TE₀₁) the value of guide wavelength will be printed out as Inf.

Note: This function is intended for interactive use. For programmatic use, see rwg_modes. The table is printed to the user's console using the PrettyTables package.

Required Positional Arguments

wgspec::AbstractString: A standard (EIA, RCSC, or IEC) abbreviation for a rectangular waveguide size. Examples include "WR650" (EIA), "WG7" (RCSC), and "R22" (IEC).

frequency: The frequency expressed as a Unitful quantity with dimensions of inverse time. Examples: 2.35u"GHz", 2350u"MHz", 2.35e9u"Hz", etc.

Optional Keyword Arguments

length_unit:: A Unitful:Units object with dimension of length. Default value is u"inch". The table will print values in the "guide wavelength" column using this unit; similarly the attenuation column will use units of dB/length_unit.

ϵᵣ or epsr: Dielectric constant of material filling the waveguide. Default value is 1.0.

tanδ or tandel: Loss tangent of material filling the waveguide. Default value is 0.0.

σ or sigma: The bulk conductivity of the waveguide metal walls expressed as a Unitful quantity with dimensions the same as those of u"S/m". Default value is Inf*u"S/m".

Rq: The RMS surface roughness of the waveguide walls expressed as a Unitful quantity having dimension of length. Default value is 0.0u"m".

col_fmts::String = ["%s", "%i", "%i", "%#.8g", "%8.4f", "%8.4f"]: A vector of C-style format strings used to format the six columns of the table.

WaveguideModes.rwg_modetable — Method

rwg_modetable(a, b, frequency; kwargs...)

An alternative method for rwg_modetable allowing explicit specification of the waveguide dimensions. Here, instead of wgspec, the first two arguments are the x and y dimensions of the waveguide, expressed as Unitful quantities having dimension of length. Both must be expressed in the same units, e.g. u"mm" or u"inch". The default value of length_unit for this method is whatever length unit a and b use.

WaveguideModes.rwgte10gz — Function

rwgte10gz(a, b, f; ϵᵣ, tanδ, σ, Rq) -> (γ, Z)

Accurately compute γ and Z (prop. constant and wave impedance) of TE10 mode in rectangular guide with rough walls.

Required Positional Arguments

a, b: The waveguide x and y dimensions, Unitful quantities with dimensions of length.

f: Frequency, a Unitful quantity with dimensions of inverse time.

Optional Keyword Arguments

ϵᵣ, tanδ: Dielectric constant and loss tangent, respectively, of material filling the waveguide. Default values are 1.0 and 0.0, respectively.

σ: The bulk conductivity of the waveguide metal walls, a Unitful quantity with the same dimensions as u"S/m". Default value is Inf*u"S/m".

Rq: The RMS surface roughness of the waveguide walls, a Unitful quantity with dimension of length. Default value is 0.0u"m"

Return Values

γ: Complex propagation constant [neper/m]

Z: Complex wave impedance [Ω]

References

[1] Lomakin, Konstantin, Gerald Gold, and Klaus Helmreich. "Analytical waveguide model precisely predicting loss and delay including surface roughness." IEEE Transactions on Microwave Theory and Techniques 66, no. 6 (2018): 2649-2662.

[2] Gold, Gerald, and Klaus Helmreich. "A physical surface roughness model and its applications." IEEE Transactions on Microwave Theory and Techniques 65, no. 10 (2017): 3720-3732.

[3] Lomakin, Konstantin, Gerald Gold, and Klaus Helmreich. "Transmission line model for rectangular waveguides accurately incorporating loss effects." In 2017 IEEE 21st Workshop on Signal and Power Integrity (SPI), pp. 1-4. IEEE, 2017.

rwgte10gz(wgspec, f; kwargs...) -> (γ, Z)

This method accepts a rectangular waveguide "specification string" instead of a and b dimensions.

Required Positional Inputs

wgspec::AbstractString: A standard (EIA, RCSC, or IEC) abbreviation for a rectangular waveguide size. Examples include "WR650" (EIA), "WG7" (RCSC), and "R22" (IEC).

Optional Keyword Arguments

ϵᵣ, tanδ: Dielectric constant and loss tangent, respectively, of material filling the waveguide. Default values are 1.0 and 0.0, respectively.

σ: The bulk conductivity of the waveguide metal walls, a Unitful quantity with the same dimensions as u"S/m". Default value is Inf*u"S/m".

Rq: The RMS surface roughness of the waveguide walls, a Unitful quantity with dimension of length. Default value is 0.0u"m"

Return Values

γ: Complex propagation constant [neper/m]

Z: Complex wave impedance [Ω]

rwgte10gz(a::Real, b::Real, f::Real; ϵᵣ::Real=1.0, tanδ::Real=0.0, σ::Real=Inf, Rq::Real=0.0) -> (γ, Z)
rwgte10gz(a::Real, b::Real, f::Real; epsr::Real=1.0, tandel::Real=0.0, sigma::Real=Inf, Rq::Real=0.0) -> (γ, Z)

This method accepts SI values as pure numbers (without attached units).

Required Positional Arguments

a, b: The waveguide x and y dimensions [m].

f: Frequency [Hz]

Optional Keyword Arguments

ϵᵣ or epsr: Dielectric constant of material filling the waveguide. Default value is 1.0.

tanδ or tandel: Loss tangent of material filling the waveguide. Default value is 0.0.

σ or sigma: The bulk conductivity of the waveguide metal walls [S/m]. Default value is Inf.

Rq: The RMS surface roughness of the waveguide walls [m]. Default value is 0.0.

Return Value

The tuple (γ, Z) where:

γ: Complex propagation constant [neper/m]

Z: Complex wave impedance [Ω]

Publicly Exported Types

WaveguideModes.RWG — Type

RWG <: HomogeneousMetallicWaveguide

A struct representing a section of metallic, rectangular waveguide, uniformly filled with dielectric. It is assumed that the (possibly non-smooth) metal walls are either perfectly conducting or form a "good conductor", in the sense that the standard perturbational formulas for losses in waveguide walls hold for higher-order modes. For the dominant TE10 and the TE01 modes, accurate formulas from the recent literature are used. See the docstring for function rwgte10gz for more information.

Fields

a::Float64: The x-dimension of the waveguide [m].

b::Float64: The y-dimension of the waveguide [m].

l::Float64: The length of the waveguide section (along z) [m].

ϵᵣ::Float64: The dielectric constant for the material filling the guide [unitless].

tanδ::Float64: The loss tangent for the material filling the guide [unitless].

σ::Float54: The bulk conductivity of the waveguide metal walls [S/m].

Rq::Float64: The metal wall RMS surface roughness [m].

modes::Vector{RWGMode}: The modes treated in this guide. If provided, the list should be sorted in order of increasing cutoff frequency.

Both positional and keyword argument constructors for RWG are available.

WaveguideModes.RWGMode — Type

RWGMode <: HomogeneousMetalPipeMode

Struct representing an electromagnetic mode of a rectangular waveguide, homogeneously filled with dielectric.

Fields:

p::TETM: Mode type.

m::Int: Index for modal variation along x direction.

n::Int: Index for modal variation along y direction.

kco::Float64: Cutoff wavenumber [radian/meter].

f::Float64: Frequency [Hz].

γ::ComplexF64: Attenuation constant [neper/meter].

Z::ComplexF64: Modal wave impedance normalized to η₀, the impedance of free space [unitless].

Non-Exported API (Subject to Change)

Private Functions

WaveguideModes.is_html_environment — Method

is_html_environment()

Returns true if code is running under IJulia or Pluto, and false otherwise.

WaveguideModes.mysqrt — Method

mysqrt(x)

Same as sqrt unless sqrt(x) is pure negative imaginary in which case it returns -sqrt(x) (i.e., positive pure imaginary).

WaveguideModes.setup_modes! — Function

setup_modes!(wg::RWG, f::Real, nmodes::Integer=length(c.modes))

Set up the modes for a uniform rectangular waveguide.

Arguments

wg: A RWG instance. If wg.modes is empty, then it will be appended to length(n). After this, or if it is already allocated, each mode in the vector will be replaced by one with updated values of γ and Z corresponding to the frequency f.

f: The frequency [Hz].

nmodes: (optional) The number of modes to append to wg.modes if it is empty. If wg.modes is nonempty, then nmodes must be equal to length(wg.modes).

WaveguideModes.update — Method

update(mode::HomogeneousMetalPipeMode; f::Real, γ::Complex, Z::Complex) -> mode

Create a copy of an existing mode with updated frequency-dependent fields and return the updated mode.

Private Types

WaveguideModes.HomogeneousMetalPipeMode — Type

abstract type HomogeneousMetalPipeMode <: MetalPipeMode end

Electromagnetic mode supported by homogeneously filled metal pipes. Concrete subtypes should contain at least the following fields:

p::TETM: Mode classification (TE or TM).

f::Float64: Frequency [Hz].

γ::ComplexF64: Frequency-dependent attenuation constant [np/m].

Z::ComplexF64: Frequency-dependent wave impedance [Ω].

WaveguideModes.MetalPipeMode — Type

abstract type MetalPipeMode <: Mode end

Electromagnetic mode supported by metal pipes, possibly filled with inhomogeneous dielectric.

WaveguideModes.Mode — Type

abstract type Mode end

Arbitrary electromagnetic mode. Should contain the following fields:

γ::ComplexF64: Attenuation constant in z direction [np/m]

This page was generated using Literate.jl.